+++ lib/guard/jasmine/coverage.rb
@@ -1,0 +1,74 @@
+# coding: utf-8
+require 'tilt'
+require 'childprocess'
+require 'guard/jasmine/util'
+
+# Tilt template to generate coverage instrumented
+# Jasmine specs files.
+#
+class JasmineCoverage < Tilt::Template
+  extend ::Guard::Jasmine::Util
+
+  def prepare
+  end
+
+  # Returns a coverage instrumented JavaScript file
+  # when the environment variable `JSCOVERAGE` is `true`
+  # and the `jscoverage` binary is in the `$PATH`.
+  #
+  def evaluate(context, locals)
+    return data unless JasmineCoverage.coverage_bin
+    return data unless is_project_asset?(file)
+
+    Dir.mktmpdir do |path|
+      filename = File.basename(file)
+      input    = File.join(path, filename).sub /\.js.+/, '.js'
+
+      File.write input, data
+
+      r, w = IO.pipe
+      proc = ChildProcess.build(JasmineCoverage.coverage_bin, 'instrument', '--embed-source', input)
+      proc.io.stdout = proc.io.stderr = w
+      proc.start
+      proc.wait
+      w.close
+
+      raise "Could not generate coverage instrumented file for #{ file }" unless proc.exit_code == 0
+
+      r.read.gsub input, file
+    end
+  end
+
+  private
+
+  # Tests if the given file is a project asset.
+  #
+  # @param [Pathname] file the path to the file
+  # @return [Boolean] true if a valid asset
+  #
+  def is_project_asset?(file)
+    file.include?(Rails.root.to_s) && ::Rails.application.config.assets.paths.find { |p| file.include?(p.to_s) }
+  end
+
+  # Returns the coverage executable path.
+  #
+  # @return [String] the path
+  #
+  def self.coverage_bin
+    @coverage_bin ||= which 'istanbul'
+  end
+
+end
+
+if ENV['COVERAGE'] == 'true' and defined?(Rails)
+
+  # Guard::Jasmine engine to register coverage instrumented
+  # Jasmine spec files.
+  #
+  class GuardJasmineCoverageEngine < ::Rails::Engine
+    config.after_initialize do |app|
+      app.assets.register_postprocessor 'application/javascript', JasmineCoverage
+    end
+  end
+
+end